requirements.txt
# API server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# HTTP client (async, used for all external calls)
httpx>=0.27.0

# Data models
pydantic>=2.9.0

# Environment / config
python-dotenv>=1.0.0

# LangChain (Phase II / III)
langchain>=0.3.0
langchain-community>=0.3.0

# YAML (user mapping config)
pyyaml>=6.0.2

# Timezone support
pytz>=2024.1

# Cache client (Redis)
redis>=5.0.0

# PostgreSQL DMS backend
asyncpg>=0.30.0

# Document Ingestion (optional: Vision OCR + file watching)
pymupdf>=1.24.0
watchfiles>=0.21.0

# Mail Ingestion
imapclient>=3.0.1
markdownify>=0.13.0

# Locale-aware date/month name parsing (used by PathTemplateParser)
babel>=2.14.0

# JSON Schema validation (prompt response validation)
jsonschema>=4.23.0

# Task queue (Phase V)
# celery>=5.4.0
